feat(package.json): update all dependencies to their latest versions for improved performance and security refactor(context.ts, jaeger-tracing.ts): change import from '@opentelemetry/tracing' to '@opentelemetry/sdk-trace-base' due to package update refactor(index.ts): move server building logic to separate file for better code organization feat(server/plugins.ts, server/server.ts): add new files to handle server plugins and server creation separately for better code modularity refactor(jaeger-setup.ts): simplify response handling in checkJaegerEndpointAvailability function for cleaner code
