+++ package.json
@@ -25,36 +25,37 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@graphql-codegen/cli": "^3.1.0",
-    "@graphql-codegen/typescript": "^3.0.1",
-    "@graphql-codegen/typescript-resolvers": "^3.1.0",
-    "@graphql-eslint/eslint-plugin": "^3.16.0",
-    "@types/node": "^18.14.0",
-    "@typescript-eslint/eslint-plugin": "^5.53.0",
-    "@typescript-eslint/parser": "^5.53.0",
+    "@graphql-codegen/cli": "^5.0.0",
+    "@graphql-codegen/typescript": "^4.0.1",
+    "@graphql-codegen/typescript-resolvers": "^4.0.1",
+    "@graphql-eslint/eslint-plugin": "^3.20.1",
+    "@types/node": "^20.5.7",
+    "@typescript-eslint/eslint-plugin": "^6.5.0",
+    "@typescript-eslint/parser": "^6.5.0",
     "@vitest/coverage-c8": "^0.28.5",
-    "artillery": "^2.0.0-30",
+    "artillery": "^2.0.0-36",
     "cross-env": "^7.0.3",
-    "eslint": "^8.34.0",
-    "prettier": "2.8.4",
+    "eslint": "^8.48.0",
+    "prettier": "3.0.3",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^5.1.6",
-    "vitest": "^0.28.5"
+    "typescript": "^5.2.2",
+    "vitest": "^0.34.3"
   },
   "dependencies": {
-    "@envelop/disable-introspection": "^4.0.6",
-    "@envelop/graphql-jit": "^5.0.6",
-    "@envelop/opentelemetry": "^4.0.6",
-    "@graphql-tools/executor-http": "^0.1.9",
-    "@graphql-tools/graphql-file-loader": "^7.5.16",
-    "@graphql-tools/load": "^7.8.13",
-    "@graphql-tools/schema": "^9.0.16",
-    "@opentelemetry/exporter-jaeger": "^1.9.1",
-    "@opentelemetry/tracing": "^0.24.0",
+    "@envelop/core": "^4.0.0",
+    "@envelop/disable-introspection": "^5.0.0",
+    "@envelop/graphql-jit": "^6.0.1",
+    "@envelop/opentelemetry": "^5.0.0",
+    "@graphql-tools/executor-http": "^1.0.2",
+    "@graphql-tools/graphql-file-loader": "^8.0.0",
+    "@graphql-tools/load": "^8.0.0",
+    "@graphql-tools/schema": "^10.0.0",
+    "@opentelemetry/exporter-jaeger": "^1.15.2",
+    "@opentelemetry/sdk-trace-base": "^1.15.2",
     "blakejs": "^1.2.1",
-    "dotenv": "^16.0.3",
-    "graphql": "^16.6.0",
-    "graphql-yoga": "3.7.0",
-    "postgres": "^3.3.3"
+    "dotenv": "^16.3.1",
+    "graphql": "^16.8.0",
+    "graphql-yoga": "4.0.4",
+    "postgres": "^3.3.5"
   }
 }
